Refactor product management: remove obsolete controller, update DbContext, and enhance product model with decimal pricing

+++ MonApiRestful/Modeles/ProduitSave/Produit.cs
@@ -8,7 +8,7 @@ namespace MonApiRestful.Modeles.ProduitSave
         public int Id { get; set; }
         public string Nom { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public double Prix { get; set; }
+        public decimal Prix { get; set; }
     }
 }
